# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VID Master Zimbabwe API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized

  # --- QUESTIONS ---
...
  /progress/leaderboard:
    get:
      operationId: getLeaderboard
      tags: [progress]
      summary: Get top learners leaderboard
      parameters:
        - in: query
          name: period
          schema:
            type: string
            enum: [daily, weekly, monthly, alltime]
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
...
  /progress/bookmarks:
    get:
      operationId: getBookmarks
      tags: [progress]
      summary: Get user's bookmarked questions
      responses:
        "200":
          description: Bookmarked questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: addBookmark
      tags: [progress]
      summary: Bookmark a question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookmarkInput"
      responses:
        "201":
          description: Bookmarked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /progress/bookmarks/{questionId}:
    delete:
      operationId: removeBookmark
      tags: [progress]
      summary: Remove a bookmark
      parameters:
        - in: path
          name: questionId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # --- ROAD SIGNS ---
...
  /admin/users:
    get:
      operationId: listUsers
      tags: [admin]
      summary: List all users (admin)
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: offset
          schema:
            type: integer
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /admin/users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [admin]
      summary: Update a user's role (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRoleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # Auth
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
        password:
          type: string
          minLength: 6
        name:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, email, name, role, xp, level, streak, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [learner, admin]
        xp:
          type: integer
        level:
          type: integer
        streak:
          type: integer
        totalTests:
          type: integer
        passRate:
          type: ["number", "null"]
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # Questions
    Question:
      type: object
      required: [id, text, options, correctAnswer, category, difficulty, explanation]
      properties:
        id:
          type: integer
        text:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: integer
        category:
          type: string
        difficulty:
          type: string
          enum: [easy, medium, hard]
        explanation:
          type: string
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, approved, published, archived]
        createdAt:
          type: string
...
    LeaderboardEntry:
      type: object
      required: [rank, userId, name, xp, level, accuracy, totalTests]
      properties:
        rank:
          type: integer
        userId:
          type: integer
        name:
          type: string
        xp:
          type: integer
        level:
          type: integer
        accuracy:
          type: number
        totalTests:
          type: integer
...
    BookmarkInput:
      type: object
      required: [questionId]
      properties:
        questionId:
          type: integer

    # Road Signs
...
    UserRoleUpdate:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [learner, admin]